[UNTESTED] Reset searchTerm when value is picked
Currently if you start writing something in the input field and then use your mouse to select one of the suggestions, the searchTerm is still showing instead of the value selected. I propose either doing like the suggested code change by resetting the field, or by setting the searchTerm field to the value of the input.